+++ lib/scss_lint/cli.rb
@@ -10,14 +10,15 @@ module SCSSLint
 
     # Subset of semantic exit codes conforming to `sysexits` documentation.
     EXIT_CODES = {
-      ok:          0,
-      warning:     1,  # One or more warnings (but no errors) were reported
-      error:       2,  # One or more errors were reported
-      usage:       64, # Command line usage error
-      no_input:    66, # Input file did not exist or was not readable
-      unavailable: 69, # Required library is unavailable
-      software:    70, # Internal software error
-      config:      78, # Configuration error
+      ok:             0,
+      warning:        1,  # One or more warnings (but no errors) were reported
+      error:          2,  # One or more errors were reported
+      usage:          64, # Command line usage error
+      no_input:       66, # Input file did not exist or was not readable
+      unavailable:    69, # Required library is unavailable
+      software:       70, # Internal software error
+      config:         78, # Configuration error
+      files_filtered: 81, # All matched files were filtered by exclusions
     }
 
     def run(args)
@@ -61,7 +62,7 @@ module SCSSLint
       end
     end
 
-    def handle_runtime_exception(exception) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+    def handle_runtime_exception(exception) # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/LineLength, Metrics/MethodLength
       case exception
       when SCSSLint::Exceptions::InvalidCLIOption
         puts exception.message
@@ -73,7 +74,10 @@ module SCSSLint
       when SCSSLint::Exceptions::RequiredLibraryMissingError
         puts exception.message
         halt :unavailable
-      when NoFilesError, Errno::ENOENT
+      when SCSSLint::Exceptions::AllFilesFilteredError
+        puts exception.message
+        halt :files_filtered
+      when SCSSLint::Exceptions::NoFilesError, Errno::ENOENT
         puts exception.message
         halt :no_input
       when NoSuchLinter