+++ spec/scss_lint/cli_spec.rb
@@ -149,5 +149,17 @@ describe SCSSLint::CLI do
         safe_run
       end
     end
+
+    context 'when all specified SCSS files are filtered by exclusions' do
+      before do
+        SCSSLint::FileFinder.any_instance.stub(:find)
+          .and_raise(SCSSLint::Exceptions::AllFilesFilteredError)
+      end
+
+      it 'exits with an appropriate status code' do
+        subject.should_receive(:halt).with(:files_filtered)
+        safe_run
+      end
+    end
   end
 end